+++ test/jsonToCsv.test.ts
@@ -52,7 +52,7 @@ describe('jsonToCsv', () => {
       assert.deepStrictEqual(
         jsonToCsv(JSON.stringify([jsonMessage])),
         'timestamp,icao24,latitude,longitude,groundspeed,track,vertical_rate,callsign,onground,alert,spi,squawk,altitude,geoaltitude,last_position,lastcontact,hour\n' +
-          `${jsonMessage.timestamp},${jsonMessage.icao24},${jsonMessage.latitude},${jsonMessage.longitude},${jsonMessage.groundspeed},${jsonMessage.track},${jsonMessage.vertical_rate},${jsonMessage.callsign},${jsonMessage.onground},${jsonMessage.alert},${jsonMessage.spi},${jsonMessage.squawk},${jsonMessage.altitude},${jsonMessage.geoaltitude},${jsonMessage.last_position},${jsonMessage.lastcontact},${jsonMessage.hour}\n`,
+          `${jsonMessage.timestamp},${jsonMessage.icao24},${jsonMessage.latitude},${jsonMessage.longitude},${jsonMessage.groundspeed},${jsonMessage.track},${jsonMessage.vertical_rate},${jsonMessage.callsign},${jsonMessage.onground},${jsonMessage.alert},${jsonMessage.spi},${jsonMessage.squawk},${jsonMessage.altitude},${jsonMessage.geoaltitude},${jsonMessage.last_position},${jsonMessage.lastcontact},${jsonMessage.hour}`,
       )
     })
 
@@ -64,7 +64,7 @@ describe('jsonToCsv', () => {
           ]),
         ),
         'timestamp,icao24,latitude,longitude,groundspeed,track,vertical_rate,callsign,onground,alert,spi,squawk,altitude,geoaltitude,last_position,lastcontact,hour\n' +
-          `${jsonMessage.timestamp},${jsonMessage.icao24},,,,,,,,,,,,,,,\n`,
+          `${jsonMessage.timestamp},${jsonMessage.icao24},,,,,,,,,,,,,,,`,
       )
     })
 
@@ -76,7 +76,7 @@ describe('jsonToCsv', () => {
           ]),
           true,
         ),
-        'timestamp,icao24,latitude,longitude,groundspeed,track,vertical_rate,callsign,onground,alert,spi,squawk,altitude,geoaltitude,last_position,lastcontact,hour,extraFields\n' +
+        'timestamp,icao24,latitude,longitude,groundspeed,track,vertical_rate,callsign,onground,alert,spi,squawk,altitude,geoaltitude,last_position,lastcontact,hour,extraField\n' +
           `${jsonMessage.timestamp},${jsonMessage.icao24},${
             jsonMessage.latitude
           },${jsonMessage.longitude},${jsonMessage.groundspeed},${
@@ -90,7 +90,7 @@ describe('jsonToCsv', () => {
           },${jsonMessage.hour},"${JSON.stringify({
             msgType: '3',
             transmissionType: 'MSG',
-          })}"\n`,
+          })}"`,
       )
     })
   })